COMMERCE-4870 commerce-machine-learning-impl Fix dispatch package name
