+++ modules/dxp/apps/commerce/commerce-machine-learning-impl/src/main/java/com/liferay/commerce/machine/learning/internal/recommendation/dispatch/executor/AnalyticsDownloadProductContentCommerceMLRecommendDispatchTaskExecutor.java
@@ -12,15 +12,15 @@
  *
  */
 
-package com.liferay.commerce.machine.learning.internal.recommendation.dispatch;
+package com.liferay.commerce.machine.learning.internal.recommendation.dispatch.executor;
 
 import com.liferay.commerce.machine.learning.internal.batch.BatchEngineTaskItemDelegateResourceMapper;
-import com.liferay.commerce.machine.learning.internal.dispatch.AnalyticsDispatchTaskExecutor;
+import com.liferay.commerce.machine.learning.internal.dispatch.executor.AnalyticsDispatchTaskExecutor;
 import com.liferay.dispatch.executor.BaseDispatchTaskExecutor;
 import com.liferay.dispatch.executor.DispatchTaskExecutor;
 import com.liferay.dispatch.executor.DispatchTaskExecutorOutput;
 import com.liferay.dispatch.model.DispatchTrigger;
-import com.liferay.headless.commerce.machine.learning.dto.v1_0.FrequentPatternRecommendation;
+import com.liferay.headless.commerce.machine.learning.dto.v1_0.ProductContentRecommendation;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.util.HashMapBuilder;
 
@@ -35,17 +35,17 @@ import org.osgi.service.component.annotations.Reference;
 @Component(
 	enabled = false, immediate = true,
 	property = {
-		"dispatch.task.executor.name=" + AnalyticsDownloadFrequentPatternCommerceMLRecommendationDispatchTaskExecutor.KEY,
-		"dispatch.task.executor.type=" + AnalyticsDownloadFrequentPatternCommerceMLRecommendationDispatchTaskExecutor.KEY
+		"dispatch.task.executor.name=" + AnalyticsDownloadProductContentCommerceMLRecommendDispatchTaskExecutor.KEY,
+		"dispatch.task.executor.type=" + AnalyticsDownloadProductContentCommerceMLRecommendDispatchTaskExecutor.KEY
 	},
 	service = DispatchTaskExecutor.class
 )
 public class
-	AnalyticsDownloadFrequentPatternCommerceMLRecommendationDispatchTaskExecutor
+	AnalyticsDownloadProductContentCommerceMLRecommendDispatchTaskExecutor
 		extends BaseDispatchTaskExecutor {
 
 	public static final String KEY =
-		"analytics-download-frequent-pattern-commerce-ml-recommendation";
+		"analytics-download-product-content-commerce-ml-recommendation";
 
 	@Override
 	public void doExecute(
@@ -56,15 +56,15 @@ public class
 		BatchEngineTaskItemDelegateResourceMapper
 			batchEngineTaskItemDelegateResourceMapper =
 				new BatchEngineTaskItemDelegateResourceMapper(
-					FrequentPatternRecommendation.class.getName(),
+					ProductContentRecommendation.class.getName(),
 					HashMapBuilder.put(
-						"antecedentIds", "antecedentIds"
-					).put(
-						"antecedentIdsLength", "antecedentIdsLength"
-					).put(
 						"createDate", "createDate"
 					).put(
+						"entryClassPK", "productId"
+					).put(
 						"jobId", "jobId"
+					).put(
+						"rank", "rank"
 					).put(
 						"recommendedEntryClassPK", "recommendedProductId"
 					).put(
